revert #9064 for now to fix #9188, #9175, #9176 (#9194)

+++ OMCompiler/SimulationRuntime/ParModelica/explicit/openclrt/OCLRuntimeUtil.cl
@@ -87,7 +87,7 @@
 #define cos(v,m) (cos(v))
 
 #ifdef DOUBLE_PREC_AVAILABLE
-typedef int  modelica_integer;
+typedef long  modelica_integer;
 typedef double modelica_real;
 #else
 typedef float modelica_real;
